tweak preview
