+++ styles/PDFPreview.module.css
@@ -11,21 +11,21 @@
     border: 1px solid black;
   }
 
-  .controls {
-      position: absolute;
-      display: flex;
-      flex-direction: row;
-      align-items: center;
-      justify-content: center;
-      bottom: 0;
-      left: 0;
-      right: 0;
-      margin-left: auto;
-      margin-right: auto;
-      padding: 4px;
-      width: 200px;
+.controls {
+    position: absolute;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    margin-left: auto;
+    margin-right: auto;
+    padding: 4px;
+    width: 200px;
 
-  }
+}
 
 .previewText {
     border: 1px solid black;
@@ -42,6 +42,10 @@
     width: 200px;
     font-size: 22px;
     background-color: white;
+}
+
+.hidden {
+    display: none;
 }
 
 .outer {
