+++ styles/PDFsDisplayEntry.module.css
@@ -31,7 +31,7 @@
     padding: 5px;
     margin: auto auto;
     padding: 2px 2px;
-    flex: 3 1 0;
+    flex: 2 1 0;
     margin: 12px;
   }
